-Edit sale form complete -Added confirm on delete option in all tables -autofocus on first field in all forms -Navigation update -Added Tooltips in all icons in table

+++ src/modules/catalog/attribute-family/attribute-family.component.ts
@@ -58,12 +58,14 @@ export class AttributeFamilyComponent implements OnInit {
 
   // For deleting attribute family
   deleteRow(id: number) {
-    this.attributeFamilyService.deleteFamily(id).subscribe(data => {
-      this.getFamily();
-      this.toast.success('Success', 'Deleted Successfully.')
-    }, err => {
-      this.toast.error('Error', 'Server error.')
-    });
+    if (confirm('Are you sure you want to delete?')) {
+      this.attributeFamilyService.deleteFamily(id).subscribe(data => {
+        this.getFamily();
+        this.toast.success('Success', 'Deleted Successfully.')
+      }, err => {
+        this.toast.error('Error', 'Server error.')
+      });
+    }
   }
 
   // For searching data from table
