+++ src/modules/catalog/edit-attribute-family/edit-attribute-family.component.ts
@@ -149,13 +149,15 @@ export class EditAttributeFamilyComponent implements OnInit {
 
   // For deleting attribute from group
   deleteRow(id: number) {
-    this.attribute.deleteAttribute(id).subscribe(data => {
-      this.getAttributesGroupData();
-      this.toast.success('Success', 'Deleted successfully.')
-    }, err => {
-      this.toast.error('Error', 'Server error.')
-    })
-    console.log('Deleted!');
+    if (confirm('Are you sure you want to delete?')) {
+      this.attribute.deleteAttribute(id).subscribe(data => {
+        this.getAttributesGroupData();
+        this.toast.success('Success', 'Deleted successfully.')
+      }, err => {
+        this.toast.error('Error', 'Server error.')
+      })
+      console.log('Deleted!');
+    }
   }
 
   // For deleting group
