+++ src/modules/item-groups/item-groups/item-groups.component.ts
@@ -55,13 +55,14 @@ export class ItemGroupsComponent implements OnInit {
 
   // For deleting Item group data
   deleteRow(id: number) {
-
-    this.itemGroupService.deleteItemGrpup(id).subscribe(data => {
-      this.getItemGroupsData();
-      this.toast.success('Success', 'Deleted Successfully.')
-    }, err => {
-      this.toast.error('Error', 'Server error.')
-    });
+    if (confirm('Are you sure you want to delete?')) {
+      this.itemGroupService.deleteItemGrpup(id).subscribe(data => {
+        this.getItemGroupsData();
+        this.toast.success('Success', 'Deleted Successfully.')
+      }, err => {
+        this.toast.error('Error', 'Server error.')
+      });
+    }
   }
 
 }
