+++ src/modules/pos/vendors/vendors.component.ts
@@ -57,13 +57,14 @@ export class VendorsComponent implements OnInit {
   }
 
   deleteRow(id: number) {
-
-    this.vendorService.deleteVendor(id).subscribe(data => {
-      this.getVendorsData();
-      this.toast.success('Success', 'Deleted Successfully.')
-    }, err => {
-      this.toast.error('Error', 'Server error.')
-    });
+    if (confirm('Are you sure you want to delete?')) {
+      this.vendorService.deleteVendor(id).subscribe(data => {
+        this.getVendorsData();
+        this.toast.success('Success', 'Deleted Successfully.')
+      }, err => {
+        this.toast.error('Error', 'Server error.')
+      });
+    }
   }
 
 }
